D110: 200+ project improvements — MODULE_VERSION, global init, macro hygiene, __user, header guards
Batch A: MODULE_VERSION("1.0") added to 26 files with MODULE_LICENSE
Batch B: Explicit =0 init for 57 static int globals across 40 files
Batch C: Macro param parenthesization (module.h, waitqueue.h) + kernel_pch.h guard fix + __user macro + uaccess __user annotations (7 items)

~73 files changed
Zero-warning build with -Wall -Wextra -Werror

+++ src/drivers/aio.c
@@ -12,15 +12,15 @@ void aio_init(void) {
     memset(aio_queue, 0, sizeof(aio_queue));
     kprintf("[OK] AIO subsystem initialized\n");
 }
-int aio_read(int fd, void *buf, size_t count, uint64_t offset) {
+ssize_t aio_read(int fd, void *buf, size_t count, uint64_t offset) {
     if (!buf) return -1;
     kprintf("[aio] read fd=%d count=%llu offset=%llu\n", fd, (unsigned long long)count, (unsigned long long)offset);
-    return (int)count;
+    return (ssize_t)count;
 }
-int aio_write(int fd, const void *buf, size_t count, uint64_t offset) {
+ssize_t aio_write(int fd, const void *buf, size_t count, uint64_t offset) {
     if (!buf) return -1;
     kprintf("[aio] write fd=%d count=%llu offset=%llu\n", fd, (unsigned long long)count, (unsigned long long)offset);
-    return (int)count;
+    return (ssize_t)count;
 }
 
 /* ── Stub: aio_fsync ─────────────────────────────── */
